Remove reducing of matched rows and fix cost's calculation
Previously if the number of rows was more than a half of the table this
number was reduced to the exact half. In some cases, if there are
a lot of matched rows (f.e. 90%) it produced incorrect low cost
of index access and more effective FTS was not used.

Removed the described reducing so index access cost is correct one
and optimizer uses FTS when it is effective.